moved validator.go to util package
